Rename folder in components/
