+++ src/ui/views/components/now-playing/NowPlaying.stories.tsx
