+++ src/ui/views/components/now-playing/NowPlaying.tsx
