+++ src/ui/views/components/now-playing/NowPlayingStatus.tsx
