+++ src/ui/views/components/now-playing/NowPlayingTitle.tsx
