+++ src/ui/views/components/sidebar-tab/SidebarTab.css
