+++ src/ui/views/session/sidebar/RoomListView.tsx
@@ -4,12 +4,12 @@ import { Room } from "@thirdroom/hydrogen-view-sdk";
 import "./RoomListView.css";
 import { Scroll } from "../../../atoms/scroll/Scroll";
 import { Avatar } from "../../../atoms/avatar/Avatar";
-import { RoomTile } from "../../components/roomtile/RoomTile";
-import { RoomTileTitle } from "../../components/roomtile/RoomTileTitle";
+import { RoomTile } from "../../components/room-tile/RoomTile";
+import { RoomTileTitle } from "../../components/room-tile/RoomTileTitle";
 import { getAvatarHttpUrl, getIdentifierColorNumber } from "../../../utils/avatar";
 import { useHydrogen } from "../../../hooks/useHydrogen";
 import { IconButton } from "../../../atoms/button/IconButton";
-import { RoomListTab } from "../../components/roomlisttab/RoomListTab";
+import { RoomListTab } from "../../components/room-list-tab/RoomListTab";
 import { CategoryHeader } from "../../components/category/CategoryHeader";
 import AddIC from "../../../../../res/ic/add.svg";
 import HomeIC from "../../../../../res/ic/home.svg";
